REFACTOR : 패키징 개선
- 외부 서버에서 응답을 받아오는 객체들로서 application 계층보다는 interface 계층에 있는 것이 바람직하다고 판단
