+++ src/main/java/org/sopt/app/common/config/ClientConfig.java
@@ -8,8 +8,8 @@ import feign.jackson.JacksonDecoder;
 import feign.jackson.JacksonEncoder;
 import feign.okhttp.OkHttpClient;
 import feign.slf4j.Slf4jLogger;
-import org.sopt.app.application.auth.client.PlaygroundClient;
-import org.sopt.app.application.operation.client.OperationClient;
+import org.sopt.app.interfaces.external.PlaygroundClient;
+import org.sopt.app.interfaces.external.OperationClient;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
